Struct re_log_types::data_cell::DataCell

pub struct DataCell {
    pub inner: Arc<DataCellInner>,
}

A cell’s worth of data, i.e. a uniform array of values for a given component type. This is the leaf type in our data model.

A DataCell can be constructed from either an iterable of native Components or directly from a slice of arrow data.

Behind the scenes, a DataCell is backed by an erased arrow array living on the heap, which is likely to point into a larger batch of contiguous memory that it shares with its peers. Cloning a DataCell is thus cheap (shallow, ref-counted).

Layout

A cell is an array of component instances: [C, C, C, …].

Consider this example:

let points: &[MyPoint] = &[[10.0, 10.0].into(), [20.0, 20.0].into(), [30.0, 30.0].into()];
let cell = DataCell::from(points);
// Or, alternatively:
let cell = DataCell::from_component::<MyPoint>([[10.0, 10.0], [20.0, 20.0], [30.0, 30.0]]);

The cell’s datatype is now a StructArray:

Struct([
   Field { name: "x", data_type: Float32, is_nullable: false, metadata: {} },
   Field { name: "y", data_type: Float32, is_nullable: false, metadata: {} },
])

Or, visualized as a cell within a larger table:

┌──────────────────────────────────────────────────┐
│ rerun.components.Point2D                                    │
╞══════════════════════════════════════════════════╡
│ [{x: 10, y: 10}, {x: 20, y: 20}, {x: 30, y: 30}] │
└──────────────────────────────────────────────────┘

Example

let points: &[MyPoint] = &[
    MyPoint { x: 10.0, y: 10.0 },
    MyPoint { x: 20.0, y: 20.0 },
    MyPoint { x: 30.0, y: 30.0 },
];
let _cell = DataCell::from(points);

// Or, alternatively:
let cell = DataCell::from_component::<MyPoint>([
    MyPoint { x: 10.0, y: 10.0 },
    MyPoint { x: 20.0, y: 20.0 },
    MyPoint { x: 30.0, y: 30.0 },
]);

eprintln!("{:#?}", cell.datatype());
eprintln!("{cell}");

Fields
inner: Arc<DataCellInner>

While the arrow data is already refcounted, the contents of the DataCell still have to be wrapped in an Arc to work around performance issues in arrow2.

See DataCellInner for more information.

Implementations

impl DataCell

pub fn try_from_native<'a, C>( values: impl IntoIterator<Item = impl Into<Cow<'a, C>>> ) -> DataCellResult<Self>where C: Component + Clone + 'a,

Builds a new DataCell from a uniform iterable of native component values.

Fails if the given iterable cannot be serialized to arrow, which should never happen when using Rerun’s built-in components.

pub fn try_from_native_sparse<'a, C>( values: impl IntoIterator<Item = Option<impl Into<Cow<'a, C>>>> ) -> DataCellResult<Self>where C: Component + Clone + 'a,

Builds a new DataCell from a uniform iterable of native component values.

Fails if the given iterable cannot be serialized to arrow, which should never happen when using Rerun’s built-in components.

pub fn from_native<'a, C>( values: impl IntoIterator<Item = impl Into<Cow<'a, C>>> ) -> Selfwhere C: Component + Clone + 'a,

Builds a new DataCell from a uniform iterable of native component values.

Panics if the given iterable cannot be serialized to arrow, which should never happen when using Rerun’s built-in components. See Self::try_from_native for the fallible alternative.

pub fn from_native_sparse<'a, C>( values: impl IntoIterator<Item = Option<impl Into<Cow<'a, C>>>> ) -> Selfwhere C: Component + Clone + 'a,

Builds a new DataCell from a uniform iterable of native component values.

Panics if the given iterable cannot be serialized to arrow, which should never happen when using Rerun’s built-in components. See Self::try_from_native for the fallible alternative.

pub fn from_component<'a, C>( values: impl IntoIterator<Item = impl Into<C>> ) -> Selfwhere C: Component + Clone + 'a + Into<Cow<'a, C>>,

Builds a cell from an iterable of items that can be turned into a Component.

pub fn from_component_sparse<'a, C>( values: impl IntoIterator<Item = Option<impl Into<C>>> ) -> Selfwhere C: Component + Clone + 'a + Into<Cow<'a, C>>,

Builds a cell from an iterable of items that can be turned into a Component.

⚠ Due to quirks in arrow2-convert, this requires consuming and collecting the passed-in iterator into a vector first. Prefer Self::from_native when performance matters.

pub fn try_from_arrow( name: ComponentName, values: Box<dyn Array> ) -> DataCellResult<Self>

Builds a new DataCell from an arrow array.

Fails if the array is not a valid list of components.

pub fn from_arrow(name: ComponentName, values: Box<dyn Array>) -> Self

Builds a new DataCell from an arrow array.

Panics if the array is not a valid list of components. See Self::try_from_arrow for the fallible alternative.

pub fn from_native_empty<C: Component>() -> Self

Builds an empty DataCell from a native component type.

pub fn try_from_arrow_empty( name: ComponentName, datatype: DataType ) -> DataCellResult<Self>

Builds an empty DataCell from an arrow datatype.

Fails if the datatype is not a valid component type.

pub fn from_arrow_empty(name: ComponentName, datatype: DataType) -> Self

Builds an empty DataCell from an arrow datatype.

Panics if the datatype is not a valid component type. See Self::try_from_arrow_empty for a fallible alternative.

pub fn to_arrow(&self) -> Box<dyn Array>

Returns the contents of the cell as an arrow array (shallow clone).

Avoid using raw arrow arrays unless you absolutely have to: prefer working directly with DataCells, DataRows & DataTables instead. If you do use them, try to keep the scope as short as possible: holding on to a raw array might prevent the datastore from releasing memory from garbage collected data.

pub fn as_arrow_ref(&self) -> &dyn Array

Returns the contents of the cell as a reference to an arrow array.

Avoid using raw arrow arrays unless you absolutely have to: prefer working directly with DataCells, DataRows & DataTables instead. If you do use them, try to keep the scope as short as possible: holding on to a raw array might prevent the datastore from releasing memory from garbage collected data.

pub fn try_to_native<'a, C: Component + 'a>(&'a self) -> DataCellResult<Vec<C>>

Returns the contents of the cell as an iterator of native components.

Fails if the underlying arrow data cannot be deserialized into C.

pub fn try_to_native_mono<'a, C: Component + 'a>( &'a self ) -> DataCellResult<Option<C>>

Returns the contents of an expected mono-component as an Option<C>.

Fails if the underlying arrow data cannot be deserialized into C.

pub fn to_native<'a, C: Component + 'a>(&'a self) -> Vec<C>

Returns the contents of the cell as an iterator of native components.

Panics if the underlying arrow data cannot be deserialized into C. See Self::try_to_native for a fallible alternative.

pub fn try_to_native_opt<'a, C: Component + 'a>( &'a self ) -> DataCellResult<Vec<Option<C>>>

Returns the contents of the cell as an iterator of native optional components.

Fails if the underlying arrow data cannot be deserialized into C.

pub fn to_native_opt<'a, C: Component + 'a>(&'a self) -> Vec<Option<C>>

Returns the contents of the cell as an iterator of native optional components.

Panics if the underlying arrow data cannot be deserialized into C. See Self::try_to_native_opt for a fallible alternative.

impl DataCell

pub fn component_name(&self) -> ComponentName

The name of the component type stored in the cell.

pub fn datatype(&self) -> &DataType

The type of the component stored in the cell, i.e. the cell is an array of that type.

pub fn num_instances(&self) -> u32

The length of the cell’s array, i.e. how many component instances are in the cell?

pub fn is_empty(&self) -> bool

pub fn is_dense(&self) -> bool

Returns true if the underlying array is dense (no nulls).

pub fn is_sorted_and_unique(&self) -> DataCellResult<bool>

Returns true if the underlying array is both sorted (increasing order) and contains only unique values.

The cell must be dense, otherwise the result of this method is undefined.

impl DataCell

pub fn compute_size_bytes(&mut self) -> bool

Compute and cache the total size (stack + heap) of the inner cell and its underlying arrow array, in bytes. This does nothing if the size has already been computed and cached before.

The caller must the sole owner of this cell, as this requires mutating an Arc under the hood. Returns false otherwise.

Beware: this is very costly!