Module re_log_types::data_table

Structs

DataCellColumn

A column’s worth of DataCells: a sparse collection of DataCells that share the same underlying type and likely point to shared, contiguous memory.

DataTable

A sparse table’s worth of data, i.e. a batch of events: a collection of DataRows. This is the top-level layer in our data model.

TableId

A unique ID for a DataTable.

Enums

DataTableError

Constants

COLUMN_ENTITY_PATH

COLUMN_INSERT_ID

COLUMN_NUM_INSTANCES

COLUMN_ROW_ID

COLUMN_TIMEPOINT

METADATA_KIND

METADATA_KIND_CONTROL

METADATA_KIND_DATA

METADATA_KIND_TIME

METADATA_TABLE_ID

Statics

ONCE 🔒

ONCE 🔒

__SEEN_MESSAGES 🔒

__SEEN_MESSAGES 🔒

Functions

cell_to_bytes 🔒

compute_rows 🔒

Given a DataTable, returns all of its rows sorted by timeline.

data_to_lists 🔒

Create a list-array out of a flattened array of cell values.

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

f 🔒

serialize_time_column 🔒

unit_values_to_unit_lists 🔒

Transforms an array of unit values into a list of unit arrays.

Type Definitions

DataCellOptVec

DataTableResult

EntityPathVec

ErasedTimeVec

NumInstancesVec

RowIdVec

TimeOptVec

TimePointVec