Struct re_log_types::data_table_batcher::DataTableBatcherInner

struct DataTableBatcherInner {
    tx_cmds: Sender<Command>,
    rx_tables: Option<Receiver<DataTable>>,
    cmds_to_tables_handle: Option<JoinHandle<()>>,
}

Fields
tx_cmds: Sender<Command>

The one and only entrypoint into the pipeline: this is never cloned nor publicly exposed, therefore the Drop implementation is guaranteed that no more data can come in while it’s running.

rx_tables: Option<Receiver<DataTable>>cmds_to_tables_handle: Option<JoinHandle<()>>
Implementations

impl DataTableBatcherInner

fn push_row(&self, row: DataRow)

fn flush_async(&self)

fn flush_blocking(&self)

fn send_cmd(&self, cmd: Command)