Module re_log_types::path

Every logged entity in Rerun is logged to an EntityPath.

The path is made up out of several EntityPathParts, each of which is either a name (EntityPathPart::Name) or an Index.

The Indexes are for tables, arrays etc.

Modules

component_name 🔒

component_path 🔒

entity_path 🔒

entity_path_impl 🔒

parse_path 🔒

Structs

ComponentName

The name of an entity component, e.g. pos or color.

ComponentPath

A EntityPath plus a ComponentName.

EntityPath

camera / "left" / points / #42

EntityPathHash

A 64 bit hash of EntityPath with very small risk of collision.

EntityPathImpl

camera / "left" / points / #42

Enums

EntityPathPart

The different parts that make up an EntityPath.

PathParseError

Functions

parse_entity_path

Parses an entity path, e.g. foo/bar/#1234/5678/"string index"/a6a5e96c-fd52-4d21-a394-ffbb6e5def1d