Struct re_log_types::path::entity_path::EntityPath

pub struct EntityPath {
    hash: EntityPathHash,
    path: Arc<EntityPathImpl>,
}

The unique identifier of an entity, e.g. camera/"ACME Örnöga"/points

The entity path is a list of parts separated by slashes.

Each part is either a name of a limited set of characters, or an Index. Names are like idenitifers in code, and must match the regex: [a-zA-z0-9_-]+ Indices are like array indices or keys in a map or table, and can be any string, uuid, or number.

Reference-counted internally, so this is cheap to clone.

Implements [nohash_hasher::IsEnabled].

assert_eq!(
    EntityPath::parse_strict(r#"camera/"ACME Örnöga"/points/#42"#).unwrap(),
    EntityPath::new(vec![
        EntityPathPart::Name("camera".into()),
        EntityPathPart::Index(Index::String("ACME Örnöga".into())),
        EntityPathPart::Name("points".into()),
        EntityPathPart::Index(Index::Sequence(42))
    ])
);

assert_eq!(
    EntityPath::data_type(),
    DataType::Extension("rerun.entity_path".into(), Box::new(DataType::Utf8), None),
);

Fields
hash: EntityPathHash

precomputed hash

path: Arc<EntityPathImpl>
Implementations

impl EntityPath

pub fn root() -> Self

pub fn new(parts: Vec<EntityPathPart>) -> Self

pub fn from_file_path_as_single_string(file_path: &Path) -> Self

Treat the file path as one opaque string.

The file path separators will NOT become splits in the new path. The returned path will only have one part.

pub fn from_single_string(string: String) -> Self

Treat the string as one opaque string, NOT splitting on any slashes.

pub fn iter(&self) -> impl Iterator<Item = &EntityPathPart>

pub fn last(&self) -> Option<&EntityPathPart>

pub fn as_slice(&self) -> &[EntityPathPart]

pub fn to_vec(&self) -> Vec<EntityPathPart>

pub fn is_root(&self) -> bool

pub fn is_descendant_of(&self, other: &EntityPath) -> bool

Is this a strict descendant of the given path.

pub fn is_child_of(&self, other: &EntityPath) -> bool

Is this a direct child of the other path.

pub fn len(&self) -> usize

Number of parts

pub fn hash(&self) -> EntityPathHash

pub fn hash64(&self) -> u64

Precomputed 64-bit hash.

pub fn parent(&self) -> Option<Self>

Return None if root.

pub fn join(&self, other: &Self) -> Self

impl EntityPath

Entity path parsing

When parsing a DataPath, it is important that we can distinguish the component and index from the actual entity path. This requires us to forbid certain characters in an entity part name. For instance, in foo/bar.baz, is baz a component name, or part of the entity path? So, when parsing a full DataPaths we are quite strict with what we allow. But when parsing EntityPaths we want to be a bit more forgiving, so we can accept things like foo/bar.baz and transform it into foo/"bar.baz". This allows user to do things like log(f"foo/{filename}", my_mesh) without Rerun throwing a fit.

pub fn parse_strict(s: &str) -> Result<Self, PathParseError>

Parse an entity path from a string, with strict checks for correctness.

Parses anything that ent_path.to_string() outputs.

For a forgiving parse that accepts anything, use Self::parse_forgiving.

pub fn parse_forgiving(s: &str) -> Self

Parses an entity path, handling any malformed input with a logged warning.

For a strict parses, use Self::parse_strict instead.