Struct re_log_types::path::EntityPath

pub struct EntityPath {
    hash: EntityPathHash,
    path: Arc<EntityPathImpl>,
}

camera / "left" / points / #42

Cheap to clone.

Implements [nohash_hasher::IsEnabled].

assert_eq!(
    EntityPath::data_type(),
    DataType::Extension("rerun.entity_path".into(), Box::new(DataType::Utf8), None),
);

Fields
hash: EntityPathHash

precomputed hash

path: Arc<EntityPathImpl>
Implementations

impl EntityPath

pub fn root() -> Self

pub fn new(parts: Vec<EntityPathPart>) -> Self

pub fn iter(&self) -> impl Iterator<Item = &EntityPathPart>

pub fn last(&self) -> Option<&EntityPathPart>

pub fn as_slice(&self) -> &[EntityPathPart]

pub fn is_root(&self) -> bool

pub fn is_descendant_of(&self, other: &EntityPath) -> bool

Is this a strict descendant of the given path.

pub fn is_child_of(&self, other: &EntityPath) -> bool

Is this a direct child of the other path.

pub fn len(&self) -> usize

Number of parts

pub fn hash(&self) -> EntityPathHash

pub fn hash64(&self) -> u64

Precomputed 64-bit hash.

pub fn parent(&self) -> Option<Self>

Return None if root.

pub fn join(&self, other: &Self) -> Self