Struct re_log_types::time_point::TimePoint

pub struct TimePoint(BTreeMap<Timeline, TimeInt>);

A point in time on any number of Timelines.

It can be represented by Time, a sequence index, or a mix of several things.

If this is empty, the data is timeless. Timeless data will show up on all timelines, past and future, and will hit all time queries. In other words, it is always there.

Tuple Fields
0: BTreeMap<Timeline, TimeInt>
Implementations

impl TimePoint

pub fn timeless() -> Self

Logging to this time means the data will show upp in all timelines, past and future. The time will be TimeInt::BEGINNING, meaning it will always be in range for any time query.

pub fn get(&self, timeline: &Timeline) -> Option<&TimeInt>

pub fn insert(&mut self, timeline: Timeline, time: TimeInt) -> Option<TimeInt>

pub fn remove(&mut self, timeline: &Timeline) -> Option<TimeInt>

pub fn is_timeless(&self) -> bool

pub fn is_empty(&self) -> bool

pub fn timelines(&self) -> impl ExactSizeIterator<Item = &Timeline>

pub fn times(&self) -> impl ExactSizeIterator<Item = &TimeInt>

pub fn iter(&self) -> impl ExactSizeIterator<Item = (&Timeline, &TimeInt)>

pub fn union_max(self, rhs: &Self) -> Self

Computes the union of two TimePoints, keeping the maximum time value in case of conflicts.