Trait re_log_types::DeserializableComponent

pub trait DeserializableComponent<ArrowFieldType = Self>where
    Self: Component + ArrowDeserialize + ArrowField<Type = ArrowFieldType> + 'static,
    Self::ArrayType: ArrowArray,
    for<'b> &'b Self::ArrayType: IntoIterator,{ }

A Component that fulfils all the conditions required to be deserialized from an Arrow payload.

Note that due to the use of HRTBs in arrow2_convert traits, you will still need an extra HRTB clause when marking a type as DeserializableComponent:

where
    T: SerializableComponent,
    for<'a> &'a T::ArrayType: IntoIterator,