Trait re_log_types::vec_deque_ext::VecDequeInsertionExt

pub trait VecDequeInsertionExt<T> {
    // Required method
    fn insert_many(
        &mut self,
        index: usize,
        values: impl ExactSizeIterator<Item = T>
    );
}

Extends VecDeque with extra insertion routines.

Required Methods

fn insert_many( &mut self, index: usize, values: impl ExactSizeIterator<Item = T> )

Inserts multiple elements at index within the deque, shifting all elements with indices greater than or equal to index towards the back.

This is O(1) if index corresponds to either the start or the end of the deque. Otherwise, this means splitting the deque into two pieces then stitching them back together at both ends of the added data.

Panics if index is out of bounds.