Crate re_log_types

The different types that make up the rerun log format.

Feature flags

serde — Enable (de)serialization using serde.

Mono-components

Some components, mostly transform related ones, are “mono-components”. This means that Rerun makes assumptions that depend on this component only taking on a singular value for all instances of an Entity. Where possible, exposed APIs will force these components to be logged as a singular instance or a splat. However, it is an error with undefined behavior to manually use lower-level APIs to log a batched mono-component.

This requirement is especially apparent with transforms: Each entity must have a unique transform chain, e.g. the entity foo/bar/baz is has the transform that is the product of foo.transform * foo/bar.transform * foo/bar/baz.transform.

Re-exports

pub use self::arrow_msg::ArrowMsg;

pub use self::time_point::TimeInt;

pub use self::time_point::TimePoint;

pub use self::time_point::TimeType;

pub use self::time_point::Timeline;

pub use self::time_point::TimelineName;

pub use self::path::*;

Modules

arrow_msg

ArrowMsg is the crate::LogMsg sub-type containing an Arrow payload.

data_cell 🔒

data_row 🔒

data_table 🔒

data_table_batcher 🔒

example_components

Example components to be used for tests and docs

external

hash

index 🔒

load_file 🔒

path

Every logged entity in Rerun is logged to an EntityPath.

serde_field

time 🔒

time_point

time_range 🔒

time_real 🔒

Macros

entity_path

Build a EntityPath:

entity_path_vec

Build a Vec<EntityPathPart>:

impl_into_enum

Structs

ApplicationId

The user-chosen name of the application doing the logging.

ComponentName

The fully-qualified name of a Component, e.g. rerun.components.Position2D.

DataCell

A cell’s worth of data, i.e. a uniform array of values for a given component type. This is the leaf type in our data model.

DataCellColumn

A column’s worth of DataCells: a sparse collection of DataCells that share the same underlying type and likely point to shared, contiguous memory.

DataCellInner

The actual contents of a DataCell.

DataCellRow

A row’s worth of DataCells: a collection of independent DataCells with different underlying datatypes and pointing to different parts of the heap.

DataRow

A row’s worth of data, i.e. an event: a list of DataCells associated with an auto-generated RowId, a user-specified TimePoint and EntityPath, and an expected number of instances. This is the middle layer in our data model.

DataTable

A sparse table’s worth of data, i.e. a batch of events: a collection of DataRows. This is the top-level layer in our data model.

DataTableBatcher

Implements an asynchronous batcher that coalesces DataRows into DataTables based upon the thresholds defined in the associated DataTableBatcherConfig.

DataTableBatcherConfig

Defines the different thresholds of the associated DataTableBatcher.

Duration

A signed duration represented as nanoseconds since unix epoch

EntityPathOpMsg

An operation (like a ‘clear’) on an EntityPath.

IndexHash

A 128 bit hash of Index with negligible risk of collision.

PythonVersion

RowId

A unique ID for a DataRow.

SetStoreInfo

StoreId

A unique id per store.

StoreInfo

Information about a recording or blueprint.

TableId

A unique ID for a DataTable.

Time

A date-time represented as nanoseconds since unix epoch

TimeRange

TimeRangeF

Like TimeRange, but using TimeReal for improved precision.

TimeReal

Either nanoseconds or sequence numbers.

Enums

BatchIndex

DataCellError

DataReadError

An error that can occur because a row in the store has inconsistent columns.

DataRowError

A problem with a row of data in the store.

DataTableBatcherError

Errors that can occur when creating/manipulating a DataTableBatcher.

DataTableError

FileSource

FromFileError

Errors from data_cells_from_file_path.

Index

The key of a table, or an array index.

LogMsg

The most general log message sent from the SDK to the server.

PathOp

Operation to perform on an EntityPath, e.g. clearing all components.

StoreKind

What kind of Store this is.

StoreSource

The source of a recording or blueprint.

TimeZone

Constants

COLUMN_ENTITY_PATH

COLUMN_INSERT_ID

COLUMN_NUM_INSTANCES

COLUMN_ROW_ID

COLUMN_TIMEPOINT

METADATA_KIND

METADATA_KIND_CONTROL

METADATA_KIND_DATA

Traits

SizeBytes

Approximations of stack and heap size for both internal and external types.

Functions

build_frame_nr

Build a (Timeline, TimeInt) tuple from frame_nr suitable for inserting in a TimePoint.

build_log_time

Build a (Timeline, TimeInt) tuple from log_time suitable for inserting in a TimePoint.

data_cells_from_file_contents

data_cells_from_file_path

Read the file at the given path.

Type Definitions

DataCellOptVec

DataCellResult

DataCellVec

DataReadResult

DataRowResult

DataTableResult

EntityPathVec

ErasedTimeVec

NumInstancesVec

RowIdVec

TimePointVec